Added Simple Apache Module in C
This one was a really cool and short tutorial, showing you how to setup
a C file to be compiled and loaded into Apache. It was just a
hello-world module, but can be the basis for developing extensions of
Apache.
I can't think of any real application to this right now, but I'm sure
there a bunch of situations where extending Apache can solve problems.