Webpack Tutorial
It did not work due to the fact you have to know all the history of
every package webpack has ever used, and how it changed from version to
version and how each of these several versions use different commands
and options to get things to work. It's just a webpage, just one react
component. This is actually the second tutorial in a row I did using
this piece of shit tool called Web-SteamingLoadOfBovineManure-Pack. Just
like the first tutorial, it did not run because webpack needs some
esoteric configuration to work. Worst finicky steaming pile of garbage tool I've ever had the
displeasure of being forced to use. It represents everything that's
wrong with web delopment these days. Piece of shit.