Pascal Review - up to "Variable Types"
I decided to conduct a quick review of the Pascal programming language,
in preparation to reading Niklaus Wirth's "Project Oberon" book. The book
is actually written in Oberon, but since it's a successor to Pascal, and
Pascal is the one Wirth language I used more than a decade ago... It's a
good starting point.

I'm following TutorialsPoint tutorial. Their stuff is alaways straight to
the point, which is great 'review material'.

Also moved previous Pascal exmaples to learn/pascal/random-examples.