prolog review - loops and decision making
also added a line break (nl, new line) to the hello world example.
prolog is weird... haskell and functional programming is weird too, but
prolog is way out there, much weirder. I have to try other logical
programming languages to see if it's just prolog that we implemented in
this semi 'esolang' way.